keystone-case-study/requirements.txt
anthropic>=0.40.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.32.0
rich>=13.7.0
pydantic>=2.6.0
tenacity>=8.2.0
requests>=2.31.0
